+++ src/renderer/src/App.tsx
@@ -3,6 +3,7 @@ import { Toaster } from "react-hot-toast";
 import { ThemeProvider } from "./components/ThemeProvider";
 import { FontProvider } from "./components/FontProvider";
 import { ProfileModalProvider } from "./components/profile/ProfileModalProvider";
+import { SettingsModalProvider } from "./components/settings/SettingsModalProvider";
 import ErrorBoundary from "./components/ErrorBoundary";
 import Welcome from "./screens/Welcome/Welcome";
 import Install from "./screens/Install/Install";
@@ -226,24 +227,26 @@ function App(): React.JSX.Element {
     <ThemeProvider>
       <FontProvider>
         <ProfileModalProvider>
-          <ErrorBoundary>
-            <div className={`app${isMac ? " is-mac" : ""}`}>
-              {isMac && <div className="drag-region" />}
-              <div className="app-content">{renderScreen()}</div>
-            </div>
-            <Toaster
-              position="bottom-right"
-              reverseOrder={false}
-              toastOptions={{
-                style: {
-                  background: "var(--bg-elevated)",
-                  color: "var(--text-primary)",
-                  border: "1px solid var(--border-bright)",
-                  fontSize: 13,
-                },
-              }}
-            />
-          </ErrorBoundary>
+          <SettingsModalProvider>
+            <ErrorBoundary>
+              <div className={`app${isMac ? " is-mac" : ""}`}>
+                {isMac && <div className="drag-region" />}
+                <div className="app-content">{renderScreen()}</div>
+              </div>
+              <Toaster
+                position="bottom-right"
+                reverseOrder={false}
+                toastOptions={{
+                  style: {
+                    background: "var(--bg-elevated)",
+                    color: "var(--text-primary)",
+                    border: "1px solid var(--border-bright)",
+                    fontSize: 13,
+                  },
+                }}
+              />
+            </ErrorBoundary>
+          </SettingsModalProvider>
         </ProfileModalProvider>
       </FontProvider>
     </ThemeProvider>
