+++ src/renderer/src/components/settings/CommunityPane.tsx
@@ -1,0 +1,188 @@
+import { Coffee, ExternalLink, Globe } from "lucide-react";
+import { useI18n } from "../useI18n";
+import BrandLogo from "../common/BrandLogo";
+import xLogo from "../../assets/logos/twitter.svg";
+import { useSettings } from "./SettingsDataContext";
+import {
+  DISCORD_COMMUNITY_URL,
+  HERMES_TELEGRAM_URL,
+  HERMES_WEBSITE_URL,
+  HERMES_X_URL,
+  KOFI_SUPPORT_URL,
+} from "./settingsHelpers";
+
+/** Community + support links and the OpenClaw → Hermes migration banner. */
+export default function CommunityPane(): React.JSX.Element {
+  const { t } = useI18n();
+  const {
+    openclawFound,
+    openclawPath,
+    migrationDismissed,
+    migrating,
+    migrationLog,
+    migrationResult,
+    migrationResultType,
+    migrationLogRef,
+    handleMigrate,
+    handleDismissMigration,
+  } = useSettings();
+
+  const open = (url: string) => () => window.hermesAPI.openExternal(url);
+
+  return (
+    <div className="settings-modal-pane">
+      <p className="settings-section-intro">
+        {t("settings.communityLinksHint")}
+      </p>
+
+      <div className="settings-link-grid">
+        <button
+          type="button"
+          className="settings-link is-primary"
+          onClick={open(DISCORD_COMMUNITY_URL)}
+          title={DISCORD_COMMUNITY_URL}
+        >
+          <span className="settings-link-icon">
+            <BrandLogo provider="discord" size={18} />
+          </span>
+          <span className="settings-link-label">
+            {t("settings.linkDiscord")}
+          </span>
+          <ExternalLink size={14} className="settings-link-arrow" />
+        </button>
+
+        <button
+          type="button"
+          className="settings-link"
+          onClick={open(HERMES_WEBSITE_URL)}
+          title={HERMES_WEBSITE_URL}
+        >
+          <span className="settings-link-icon">
+            <Globe size={18} />
+          </span>
+          <span className="settings-link-label">
+            {t("settings.linkWebsite")}
+          </span>
+          <ExternalLink size={14} className="settings-link-arrow" />
+        </button>
+
+        <button
+          type="button"
+          className="settings-link"
+          onClick={open(HERMES_X_URL)}
+          title={HERMES_X_URL}
+        >
+          <span className="settings-link-icon">
+            <img
+              src={xLogo}
+              width={16}
+              height={16}
+              className="brand-logo brand-logo--match-theme"
+              alt="X"
+            />
+          </span>
+          <span className="settings-link-label">{t("settings.linkX")}</span>
+          <ExternalLink size={14} className="settings-link-arrow" />
+        </button>
+
+        <button
+          type="button"
+          className="settings-link"
+          onClick={open(HERMES_TELEGRAM_URL)}
+          title={HERMES_TELEGRAM_URL}
+        >
+          <span className="settings-link-icon">
+            <BrandLogo provider="telegram" size={18} />
+          </span>
+          <span className="settings-link-label">
+            {t("settings.linkTelegram")}
+          </span>
+          <ExternalLink size={14} className="settings-link-arrow" />
+        </button>
+      </div>
+
+      <section className="settings-card">
+        <header className="settings-card-head">
+          <span className="settings-card-icon">
+            <Coffee size={18} />
+          </span>
+          <div className="settings-card-headtext">
+            <div className="settings-card-title">
+              {t("settings.supportTitle")}
+            </div>
+            <div className="settings-card-sub">{t("settings.supportHint")}</div>
+          </div>
+        </header>
+        <div className="settings-card-body">
+          <div className="settings-card-actions">
+            <button
+              type="button"
+              className="btn btn-primary"
+              onClick={open(KOFI_SUPPORT_URL)}
+              title={KOFI_SUPPORT_URL}
+            >
+              <Coffee size={14} />
+              {t("settings.supportKofi")}
+            </button>
+          </div>
+        </div>
+      </section>
+
+      {openclawFound && !migrationDismissed && (
+        <div className="settings-migration-banner">
+          <div className="settings-migration-header">
+            <div>
+              <div className="settings-migration-title">
+                {t("settings.migrationDetected")}
+              </div>
+              <div
+                className="settings-migration-desc"
+                dangerouslySetInnerHTML={{
+                  __html: t("settings.migrationDesc", {
+                    path: openclawPath || "",
+                  }),
+                }}
+              />
+            </div>
+            <button
+              className="btn-ghost settings-migration-dismiss"
+              onClick={handleDismissMigration}
+              title={t("settings.migrationDismiss")}
+            >
+              &times;
+            </button>
+          </div>
+          {migrationLog && (
+            <pre className="settings-hermes-doctor" ref={migrationLogRef}>
+              {migrationLog}
+            </pre>
+          )}
+          {migrationResult && (
+            <div
+              className={`settings-hermes-result ${migrationResultType || "error"}`}
+            >
+              {migrationResult}
+            </div>
+          )}
+          <div className="settings-migration-actions">
+            <button
+              className="btn btn-primary "
+              onClick={handleMigrate}
+              disabled={migrating}
+            >
+              {migrating
+                ? t("settings.migrating")
+                : t("settings.migrateToHermes")}
+            </button>
+            <button
+              className="btn btn-secondary "
+              onClick={handleDismissMigration}
+            >
+              {t("settings.skip")}
+            </button>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+}
