+++ src/renderer/src/components/settings/LanguagePane.tsx
@@ -1,0 +1,88 @@
+import { useEffect, useRef, useState } from "react";
+import { Check, ChevronDown } from "lucide-react";
+import { useI18n } from "../useI18n";
+import { APP_LOCALES, type AppLocale } from "../../../../shared/i18n";
+import { LANGUAGE_NATIVE_NAMES } from "./settingsHelpers";
+
+/** Interface language selector. */
+export default function LanguagePane(): React.JSX.Element {
+  const { t, locale, setLocale } = useI18n();
+  return (
+    <div className="settings-modal-pane">
+      <div className="settings-field">
+        <label className="settings-field-label">
+          {t("settings.language.label")}
+        </label>
+        <LanguageSelect locale={locale} onSelect={setLocale} />
+        <div className="settings-field-hint">{t("settings.language.hint")}</div>
+      </div>
+    </div>
+  );
+}
+
+function LanguageSelect({
+  locale,
+  onSelect,
+}: {
+  locale: AppLocale;
+  onSelect: (l: AppLocale) => void;
+}): React.JSX.Element {
+  const [isOpen, setIsOpen] = useState(false);
+  const ref = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    if (!isOpen) return;
+    function handleClickOutside(e: MouseEvent): void {
+      if (ref.current && !ref.current.contains(e.target as Node)) {
+        setIsOpen(false);
+      }
+    }
+    function handleKey(e: KeyboardEvent): void {
+      if (e.key === "Escape") setIsOpen(false);
+    }
+    document.addEventListener("mousedown", handleClickOutside);
+    document.addEventListener("keydown", handleKey);
+    return () => {
+      document.removeEventListener("mousedown", handleClickOutside);
+      document.removeEventListener("keydown", handleKey);
+    };
+  }, [isOpen]);
+
+  return (
+    <div className="settings-language-select" ref={ref}>
+      <button
+        type="button"
+        className="settings-language-trigger"
+        onClick={() => setIsOpen((v) => !v)}
+        aria-haspopup="listbox"
+        aria-expanded={isOpen}
+      >
+        <span>{LANGUAGE_NATIVE_NAMES[locale]}</span>
+        <ChevronDown size={14} />
+      </button>
+      {isOpen && (
+        <div className="settings-language-dropdown" role="listbox">
+          {APP_LOCALES.map((l) => {
+            const active = l === locale;
+            return (
+              <button
+                key={l}
+                type="button"
+                role="option"
+                aria-selected={active}
+                className={`settings-language-option ${active ? "active" : ""}`}
+                onClick={() => {
+                  onSelect(l);
+                  setIsOpen(false);
+                }}
+              >
+                <span>{LANGUAGE_NATIVE_NAMES[l]}</span>
+                {active && <Check size={14} />}
+              </button>
+            );
+          })}
+        </div>
+      )}
+    </div>
+  );
+}
