+++ src/shared/i18n/locales/en/settings.ts
@@ -6,6 +6,32 @@ export default {
     privacy: "Privacy",
     credentialPool: "Credential Pool",
   },
+  nav: {
+    groups: {
+      general: "General",
+      hermes: "Hermes Agent",
+    },
+    appearance: "Appearance",
+    language: "Language",
+    privacy: "Privacy",
+    connection: "Connection",
+    network: "Network",
+    data: "Data",
+    about: "About & Updates",
+    community: "Community",
+    logs: "Logs & Diagnostics",
+  },
+  agentSubtitle: "The local AI engine",
+  desktopTitle: "Hermes One Desktop",
+  desktopSubtitle: "This desktop app",
+  statusUpToDate: "Up to date",
+  statusUpdateAvailable: "Update available",
+  statusUpdateReady: "Update ready",
+  checkForUpdates: "Check for updates",
+  checkingUpdates: "Checking…",
+  downloadUpdate: "Download update",
+  retry: "Retry",
+  onLatestVersion: "You're on the latest version.",
   theme: {
     label: "Theme",
     system: "System",
@@ -138,6 +164,16 @@ export default {
   communityHint:
     "Join our Discord channel to ask questions, report issues, and chat with other Hermes users.",
   joinDiscord: "Join Discord Channel",
+  communityLinksHint:
+    "Connect with the Hermes One community, get help, and stay up to date.",
+  linkWebsite: "Website",
+  linkDiscord: "Discord",
+  linkX: "X",
+  linkTelegram: "Telegram",
+  supportTitle: "Support the developer",
+  supportHint:
+    "Hermes One is free and open source — if it helps you, consider supporting its development.",
+  supportKofi: "Support on Ko-fi",
 
   // SSH & Server Config
   modeSsh: "SSH Tunnel",
